requirements.txt
Pillow==10.4.0
Pylette==3.0.1
